+++ MyUtil.h
@@ -24,7 +24,7 @@ using namespace std;
                      : \
                          ((x) >= (z) ? (x) : (z)))
 						 
-void _gnuplotInit(Gnuplot& gnuPlot, const char* title, int ymin=0, int ymax=0);
+void _gnuplotInit(Gnuplot& gnuPlot, const char* title, double ymin=0, double ymax=0);
 void _gnuplotVerticalLine(Gnuplot& gnuPlot, int x, const char* dataName="");
 void _gnuplotLED(Gnuplot& gnuPlot, int nBeginLight, int nTwoLight);
